+++ backend/src/main/java/com/miu/person/LibraryMember.java
@@ -1,5 +1,6 @@
 package com.miu.person;
 
+import com.miu.Address;
 import com.miu.book.BookCopy;
 
 import java.util.List;
